cope with null values in BINARY reader
Something similar should really be done for TABLEDATA as well,
though usually TABLEDATA uses empty TD elements for nulls.

The implementation is inefficient, since any VALUES/@null attribute
is parsed for every row; tests show it taking about an extra 50% of
time to read a BINARY/BINARY2 table.
The fieldattrs function is anyway getting called for every row,
which is not very efficient; these issues should (will) be tackled
in a future commit.