simplify and speed up TABLEDATA processing
Simplify some of the code that parses the TD elements.
I don't understand the details of what was going on before,
but it was very slow.  Following this commit, parsing TABLEDATA
is about five times faster and all the unit tests still pass.

However, I might be missing something about the previous implementation,
so there might be some good reason not to apply this commit.